refactor: rename `MonitoringCommandSet` to `HealthCheckCommandSet` and move to root package
