+++ src/main/kotlin/com/github/frozensync/HealthCheckCommandSet.kt
@@ -1,4 +1,4 @@
-package com.github.frozensync.monitoring
+package com.github.frozensync
 
 import com.github.frozensync.discord.command.Command
 import com.github.frozensync.discord.command.CommandSet
@@ -6,7 +6,7 @@ import kotlinx.coroutines.reactive.awaitFirst
 import java.lang.management.ManagementFactory
 import java.time.Duration
 
-class MonitoringCommandSet : CommandSet {
+class HealthCheckCommandSet : CommandSet {
 
     override val commands: Map<String, Command> = mutableMapOf<String, Command>().apply {
         this["ping"] = { event ->
